agent/requirements.txt
fastapi
uvicorn
python-dotenv
langchain-openai
langchain
langchain-milvus
pymilvus
PyPDF2
